Explicit casts to avoid conversion warnings
Visual Studio warns about signed/unsigned conversion and limited range of types.

+++ src/ansi-c/expr2c.cpp
@@ -2152,7 +2152,7 @@ std::string expr2ct::convert_array(const exprt &src)
       if(last_was_hex)
       {
         // we use "string splicing" to avoid ambiguity
-        if(isxdigit(ch))
+        if(isxdigit(static_cast<int>(ch)))
           dest+="\" \"";
 
         last_was_hex=false;
